Extracted compoenents from Schedule Wizard
