+++ tournament-scheduler/src/models/Registration.ts
@@ -2,3 +2,7 @@ export default interface Registration {
   id: number
   participants: string[]
 }
+
+export interface RegistrationProxy extends Registration {
+  hasChanged?: boolean
+}
